Extract HWB color space into sub-module
